+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `Clock` interface has a `sleep()` method
 - `PsrToOrisaiClockAdapter` - makes any PSR clock compatible with `Orisai\Clock\Clock`
 - `SymfonyToOrisaiClockAdapter` - makes any symfony/clock compatible with `Orisai\Clock\Clock`
+- `ClockAdapterFactory` - chooses best supported adapter for `Orisai\Clock\Clock` compatibility
 
 ### Changed